backend/python/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
Flask>=2.0.0
Flask-CORS>=3.0.0
ultralytics>=8.0.0
